libmultipath: allow force reload with no active paths
Multipath is fully capable of reloading a multipath device that has no
active paths. This has been possible for years. If multipath should
reload the device, it should do so, even if the paths are down.

Generally, when multipath is force reloaded, kpartx will be updated.
However when a device is reloaded with no paths, the udev rules won't
run kpartx.  But they also weren't running kpartx when the first valid
path appeared, even though the dm activation rules get run in this case.
This changes 11-dm-mpath.rules to run kpartx when a device goes from no
usable paths to having usable paths.