multipath: add libmpathvalid library
This library allows other programs to check if a path should be claimed
by multipath. It exports an init and exit function, that need to be
called before any other functions can be called, a pointer to a struct
config, that stores the configuration which is dealt with in the
init and exit functions, and two more functions.

mpath_get_mode() get the configured find_multipaths mode.
mpath_is_path() returns whether the device is claimed by multipath, and
optionally returns the wwid.  This code works slightly different than
the multipath -c/u code for SMART mode.  Instead of checking all the
existing paths to see if another has the same wwid, it expects the
caller to pass in an array of the already known path wwids, and checks
if the current path matches any of those.

The library also doesn't set up the device-mapper log fuctions. It
leaves this up to the caller.